revert: fix: dogfood on both push and pr
This reverts commit 9173e3c4aff3a4fc7eb747cfa0f5096d516868e6.

`$TRAVIS_PULL_REQUEST` should indicate whether the build was a
"pull request" build. When a branch is pushed to, and it also has
an open PR, Travis will trigger two builds at the same time. If we
run this plugin on both builds, there's a race condition that may
cause a duplicate changelog comment to be posted.