fix: keep from unnecessarily recreating "no release" comment
We were wiping the “no release” comment on each run, even if nothing
changed. This became especially problematic when used with
`semantic-release-monorepo`, recreating the “no release” comment for
any package without a new release (likely the vast majority of
packages).

BREAKING CHANGE:
Dropped use of `verifyConditions` as a plugin hook to delete stale
comments. This now happens in `analyzeCommits` instead.